Do anything only if update.sh.next is present

+++ update-versions.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 
-mv versions.sh.next versions.sh
+if test -e versions.sh.next; then
+  mv versions.sh.next versions.sh
 
-./generate-readmes.sh
+  ./generate-readmes.sh
+fi
